version 1.1 2013-07-01
Version 1.1, 2013-07-01, tmgriffiths

• Rewrote the \supervisor command to have an optional argument giving
the number of supervisors. I used the etoolbox package and its
\ifnumcom macros to do so. The old options nosupervisor, onesupervisor
and multiplesupervisors still exist but mainly for legacy reasons. It
is not necessary to specify them in the preamble anymore. The only
thing of note is that nosupervisor is achieved by either nosupervisor
in the preamble or \supervisor[0]{} in the document Simply leaving out
supervisor will not remove the field.

• Rewrote the way each degree type specified partial or complete
requirement in the \@requirement macro. I’m now using \if statements to
set the field.

• Implemented the \declaration command (see examples). It typesets
declaration page that PhD students are required to have in the
frontmatter of their thesis.