Version 2.0
Created seperate directories for the CTAN and UOW versions.

Started cleaning up the wording of the declaration and the title page.

Added "Doctor of Philosophy" as the degree definition if option phd (or
PhD) is givven to the package.

Added alternate capitalisation of phd (PhD) as a package option

Added two new ifs: \ifhonours and \ifmasters

Started fixing the indenting in uowthesistitlepage.sty