Version 3.0 UOW 20167 rebrand
This version is a major redesign of the title page to coincide the
re-branding of UOW during 2016. It also adds the copyright statement
and acknowledgement of the Australian Government Research Training
Partnership Scholarships. Be sure to use the new version of the UOW
crest (below). A copy is distributed with the uow thesis template on
UOW website. It follows the UOW thesis guidelines and is has a basic
setup ready to get you started with your thesis. For those starting off
on their own, the following code will give you a document that
satisfies the UOW thesis guidelines.

See the documentation for changes (yes I know this is a lazy commit
message).

This is ready for merging into the master branch and to be released to
CTAN (via that branch).